+++ ReleaseNotes.md
@@ -3,6 +3,7 @@
 - form: fix defaultVal in textarea
 - exclude dedicated pages from navigation bar or view by `...-nonav` paggeId
 - hidden field fix
+- plan HTML fix and added to [complex layout example](https://github.com/ma-ha/easy-web-app/tree/master/examples/complex-layout)
 
 ## 0.17.x
 - Table/List: support number